move service daemon into its own module
